Refaktorerer validering av vedlegg
