+++ src/main/kotlin/no/nav/helse/Endepunkter.kt
@@ -8,4 +8,4 @@ const val MELLOMLAGRING_URL = "/mellomlagring"
 const val ARBEIDSGIVER_URL = "/arbeidsgiver"
 const val VEDLEGG_MED_ID_URL = "/vedlegg/{vedleggId}"
 const val VEDLEGG_URL = "/vedlegg"
-const val VALIDER_VEDLEGG = "/soknad/valider/vedlegg"
+const val VALIDER_VEDLEGG_URL = "/vedlegg/valider"
