+++ src/main/kotlin/no/nav/helse/vedlegg/VedleggListe.kt
@@ -1,0 +1,7 @@
+package no.nav.helse.vedlegg
+
+import java.net.URL
+
+data class VedleggListe (
+    val vedleggUrl: List<URL>
+)
